requirements.txt
pypdf==3.15.0
reportlab==4.0.4
requests==2.27.1
svglib==1.5.1
beautifulsoup4==4.11.1
Pillow==9.4.0
